src/FEEDback/requirements.txt
django==4.0.1
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0